--- src/app/_components/post-body.tsx
@@ -1,16 +1,0 @@
-import markdownStyles from "./markdown-styles.module.css";
-
-type Props = {
-  content: string;
-};
-
-export function PostBody({ content }: Props) {
-  return (
-    <div className="max-w-2xl mx-auto znc">
-      <div
-        className={markdownStyles["markdown"]}
-        dangerouslySetInnerHTML={{ __html: content }}
-      />
-    </div>
-  );
-}
