+++ src/app/posts/[slug]/page.tsx
@@ -8,6 +8,7 @@ import SideColumn from "@/app/_components/_common/_side-column/side-column";
 import MainColumn from "@/app/_components/_common/_main-column/main-column";
 import IntroCard from "@/app/_components/_common/_side-column/intro-card";
 import TagsCard from "@/app/_components/_common/_side-column/tags-card";
+import { PostBody } from "@/app/_components/_posts/post-body";
 
 export default async function Post(props: Params) {
   const params = await props.params;
@@ -22,15 +23,17 @@ export default async function Post(props: Params) {
   });
 
   return (
-    <div className="md:flex md:flex-row-reverse">
-      <SideColumn>
-        {post.preview && <Alert />}
-        <IntroCard />
-        <TagsCard />
-      </SideColumn>
-      <MainColumn>
-        post
-      </MainColumn>
+    <div>
+      {post.preview && <Alert />}
+      <div className="md:flex md:flex-row-reverse">
+        <SideColumn>
+          <IntroCard />
+          <TagsCard />
+        </SideColumn>
+        <MainColumn>
+          <PostBody content={content} />
+        </MainColumn>
+      </div>
     </div>
   );
 }
